heartbeat cycle 12: apply 4 Sev 7-9 code-review fixes + integrate direction review
Direction review (TL;DR): cross-class lesion-underbite at rank-32 across
3/3 architectures is itself supportive (not dispositive) of the
architecture-prior thesis. Frame g159 verdict as INCOMPLETE/SCALE-LIMITED
with cross-class null finding interpretation. >6/10 §0.1 score.

Code review fixes:

  Sev 9 g157c microbenchmark dtype mismatch: prior 'fix' set fake_h to
    fp32 but probe stayed bf16 -> RuntimeError on first matmul.
    Fixed: probe also fp32.

  Sev 8 g160 KD cache key wrong train_seed: cache meta said train_seed=42
    but actual c4_clean_v1 uses seed=160. Filename also lacked seed
    dimension. Fixed: C4_TRAIN_SEED=160 in both meta and filename.

  Sev 7 g160 verdict logic incompatible with 1-seed PILOT: criterion
    required '>=2/3 seeds' impossible with SEEDS=[42]. Fixed: PILOT_*
    prefix on verdict labels when single-seed; canonical PASS reserved
    for 3-seed g160c follow-up.

  Sev 7 g157d benchmark dtype + n_layers: same bf16/fp32 mismatch as
    g157c, plus hardcoded n_layers_used=3 when g157d uses 5 depths.
    Fixed: float32 fake tensors and probe; n_layers_used = max len of
    MIDBAND_INDICES values (handles 5-depth scope dynamically).

g161_run was clean per cycle 12 review.

Direction review committed for audit trail.

Committed by Devansh

+++ code/genome_157c_3seed_canonical_verdict.py
@@ -332,7 +332,9 @@ def microbenchmark(hidden, vocab, n_ckpts):
         ("local", LocalMLPProbe, False),
         ("prefix_embed", PrefixEmbedAttnProbe, True),
     ]:
-        probe = ProbeClass(hidden, vocab).to("cuda").to(torch.bfloat16)
+        # Per cycle 12 code review Sev-9: probe must be FP32 to match
+        # train_probe path (mat1/mat2 dtype mismatch crashes otherwise)
+        probe = ProbeClass(hidden, vocab).to("cuda").to(torch.float32)
         opt = torch.optim.AdamW(probe.parameters(), lr=PROBE_LR)
         torch.cuda.synchronize()
         t0 = time.time()
